chore: rename modules
